Use a partial class to move ClassicUO entry point to ClassicUO.cs
Keep the classicuo code in the classicuo file.
